Intermediate update before V1 release to introduce some large changes to functionality into main branch  (#3)
Worked to adjust the methods used in the Sampler/QuickSampler so the functionality and behavior of these is more clear and useful. Now, two functions exist, a sample_N_inputs and sample_N_outputs, which replace the original sample_N_states.

Also added to adjusted documentation, both in the sphinx docs and in class docstrings.